# src/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Octavio Poultry Farm API Documentation",
    "description": "API contract and interactive documentation for the Octavio Poultry Farm management backend. Defines schemas, routes, parameters, and role-based permissions.",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "/api",
      "description": "API Base URL (relies on browser host or proxy)"
    }
  ],
  "security": [
    {
      "CookieAuth": []
    },
    {
      "BearerAuth": []
    }
  ],
...
  "paths": {
...
    "/batches": {
      "get": {
        "tags": ["Batches"],
        "summary": "List all batch cycles",
        "description": "Returns a list of all batch cycles recorded for the authenticated user's farm.",
        "responses": {
          "200": {
            "description": "List of batches",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": { "$ref": "#/components/schemas/Batch" }
                }
              }
            }
          },
          "401": { "description": "Unauthorized" }
        }
      },
      "post": {
        "tags": ["Batches"],
        "summary": "Create a new batch cycle",
        "description": "Registers a new flock cycle. Requires role: OperationManager or PrimaryOwner. Validates and initializes defaults for building-level loadings.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/BatchCreate" }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Batch created successfully",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Batch" }
              }
            }
          },
          "400": { "description": "Validation error" },
          "401": { "description": "Unauthorized" },
          "403": { "description": "Forbidden - requires OperationManager" }
        }
      }
    },
...
    "/employees": {
      "get": {
        "tags": ["Employees & Payroll"],
        "summary": "List active employees",
        "description": "Retrieves profile settings, displays, phone, addresses, and hire histories. Requires role: OperationManager or PrimaryOwner.",
        "responses": {
          "200": {
            "description": "List of active employees",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": { "$ref": "#/components/schemas/Employee" }
                }
              }
            }
          },
          "401": { "description": "Unauthorized" },
          "403": { "description": "Forbidden" }
        }
      },
      "post": {
        "tags": ["Employees & Payroll"],
        "summary": "Create an employee profile",
        "description": "Registers a new employee stakeholder profile. Requires role: OperationManager or PrimaryOwner.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/EmployeeCreate" }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Employee profile created",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Employee" }
              }
            }
          },
          "400": { "description": "Validation error" },
          "409": { "description": "Stakeholder name collision" }
        }
      }
    },
...
    "/logs": {
      "get": {
        "tags": ["Daily Logs"],
        "summary": "List daily logs",
        "description": "Retrieves daily operational reports (feed consumed, weights, mortalities). Supports filtering by batchId.",
        "parameters": [
          {
            "name": "batchId",
            "in": "query",
            "required": false,
            "schema": { "type": "string" },
            "description": "Filter by batch ID"
          }
        ],
        "responses": {
          "200": {
            "description": "List of logs",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": { "$ref": "#/components/schemas/DailyLog" }
                }
              }
            }
          },
          "401": { "description": "Unauthorized" }
        }
      },
      "post": {
        "tags": ["Daily Logs"],
        "summary": "Submit a new daily operational log",
        "description": "Registers daily mortality, weights, and feeds consumed. Generates automatic inventory write-downs. Requires role: DataEntry or above.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/DailyLogCreate" }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Daily log registered",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/DailyLog" }
              }
            }
          },
          "400": { "description": "Validation error or invalid assignment" },
          "404": { "description": "Batch or employee profile not found" }
        }
      }
    },
...
    "/transactions": {
      "get": {
        "tags": ["Transactions & Ledger"],
        "summary": "List all transactions",
        "description": "Retrieves transactions ledger records. Supports batch filtering.",
        "parameters": [
          {
            "name": "batchId",
            "in": "query",
            "required": false,
            "schema": { "type": "string" },
            "description": "Filter by batch ID"
          }
        ],
        "responses": {
          "200": {
            "description": "List of transactions",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": { "$ref": "#/components/schemas/Transaction" }
                }
              }
            }
          },
          "401": { "description": "Unauthorized" }
        }
      },
      "post": {
        "tags": ["Transactions & Ledger"],
        "summary": "Post a new cash ledger transaction",
        "description": "Inserts cash ledger entry. Requires role: OperationManager or PrimaryOwner.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/TransactionCreate" }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Transaction created",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Transaction" }
              }
            }
          },
          "400": { "description": "Validation error" }
        }
      }
    },
...
    "/public/current-batch": {
      "get": {
        "tags": ["Public / AI Chat"],
        "summary": "Retrieve active batch without logging in",
        "description": "Returns active flock code, notes, and metrics for public visibility widgets.",
        "security": [],
        "responses": {
          "200": {
            "description": "Active snapshot details",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Batch" }
              }
            }
          },
          "404": { "description": "No active batch" }
        }
      }
    },
...
  },
  "components": {
...
    "schemas": {
...
      "Batch": {
        "type": "object",
        "properties": {
          "id": { "type": "string", "example": "20260418" },
          "batchCode": { "type": "string", "example": "20260418" },
          "startDate": { "type": "string", "example": "2026-04-18" },
          "targetHarvestDate": { "type": "string", "nullable": true, "example": "2026-05-28" },
          "actualHarvestEndDate": { "type": "string", "nullable": true, "example": null },
          "status": { "type": "string", "enum": ["Active", "Completed"], "example": "Active" },
          "totalChicksLoaded": { "type": "integer", "example": 25000 },
          "actualChicksArrived": { "type": "integer", "example": 25000 },
          "doaCount": { "type": "integer", "example": 42 },
          "netChicksPlaced": { "type": "integer", "example": 24958 },
          "arrivalSampleWeightGrams": { "type": "number", "nullable": true, "example": 42.5 },
          "plannedFlock": { "type": "integer", "example": 25000 },
          "targetFeedKg": { "type": "number", "example": 50000 },
          "notes": { "type": "string", "example": "Initial setup notes." }
        }
      },
      "BatchCreate": {
        "type": "object",
        "properties": {
          "startDate": { "type": "string", "example": "2026-04-18" },
          "targetHarvestDate": { "type": "string", "nullable": true, "example": "2026-05-28" },
          "totalChicksLoaded": { "type": "integer", "example": 25000 },
          "actualChicksArrived": { "type": "integer", "example": 25000 },
          "doaCount": { "type": "integer", "example": 42 },
          "netChicksPlaced": { "type": "integer", "example": 24958 },
          "arrivalSampleWeightGrams": { "type": "number", "nullable": true, "example": 42.5 },
          "plannedFlock": { "type": "integer", "example": 25000 },
          "targetFeedKg": { "type": "number", "example": 50000 },
          "notes": { "type": "string", "nullable": true },
          "status": { "type": "string", "enum": ["Active", "Completed"] },
          "loadings": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/LoadingItem" }
          }
        },
        "required": ["startDate", "totalChicksLoaded", "plannedFlock", "targetFeedKg"]
      },
      "LoadingItem": {
        "type": "object",
        "properties": {
          "building": { "type": "string", "example": "A" },
          "buildingName": { "type": "string", "example": "A" },
          "chicksLoaded": { "type": "integer", "example": 10000 },
          "doaCount": { "type": "integer", "example": 12 },
          "netChicksPlaced": { "type": "integer", "example": 9988 },
          "sampleWeightGrams": { "type": "number", "nullable": true, "example": 42.5 },
          "loadingSharePct": { "type": "number", "example": 40 },
          "remarks": { "type": "string", "nullable": true }
        }
      },
...
      "Employee": {
        "type": "object",
        "properties": {
          "id": { "type": "integer" },
          "name": { "type": "string" },
          "displayName": { "type": "string" },
          "phone": { "type": "string", "nullable": true },
          "email": { "type": "string", "nullable": true },
          "address": { "type": "string", "nullable": true },
          "position": { "type": "string" },
          "hireDate": { "type": "string" },
          "assignedBuilding": { "type": "string" },
          "notes": { "type": "string" },
          "isActive": { "type": "boolean" }
        }
      },
      "EmployeeCreate": {
        "type": "object",
        "properties": {
          "name": { "type": "string" },
          "displayName": { "type": "string" },
          "phone": { "type": "string" },
          "email": { "type": "string" },
          "address": { "type": "string" },
          "position": { "type": "string" },
          "hireDate": { "type": "string" },
          "assignedBuilding": { "type": "string" },
          "notes": { "type": "string" }
        },
        "required": ["name"]
      },
...
      "DailyLog": {
        "type": "object",
        "properties": {
          "id": { "type": "integer" },
          "batchId": { "type": "string" },
          "date": { "type": "string" },
          "building": { "type": "string" },
          "employeeId": { "type": "integer" },
          "employeeName": { "type": "string" },
          "handledBirds": { "type": "integer" },
          "feedItemId": { "type": "integer" },
          "feedItemName": { "type": "string" },
          "feed": { "type": "number" },
          "mortality": { "type": "integer" },
          "averageWeightGrams": { "type": "number", "nullable": true },
          "remarks": { "type": "string" }
        }
      },
      "DailyLogCreate": {
        "type": "object",
        "properties": {
          "batchId": { "type": "string" },
          "date": { "type": "string" },
          "building": { "type": "string" },
          "employeeId": { "type": "integer" },
          "handledBirds": { "type": "integer" },
          "feedItemId": { "type": "integer" },
          "feed": { "type": "number" },
          "mortality": { "type": "integer" },
          "averageWeightGrams": { "type": "number" },
          "remarks": { "type": "string" }
        },
        "required": ["date", "building", "employeeId"]
      },
      "Transaction": {
        "type": "object",
        "properties": {
          "id": { "type": "integer" },
          "batchId": { "type": "string", "nullable": true },
          "date": { "type": "string" },
          "building": { "type": "string" },
          "fundingNature": { "type": "string", "enum": ["OPEX", "CAPEX", "Revenue", "Investment", "Draw"] },
          "category": { "type": "string" },
          "categoryId": { "type": "integer" },
          "description": { "type": "string" },
          "quantity": { "type": "number", "nullable": true },
          "unitCost": { "type": "number", "nullable": true },
          "amount": { "type": "number" },
          "paidBy": { "type": "integer", "nullable": true },
          "paidByName": { "type": "string" },
          "paidTo": { "type": "integer", "nullable": true },
          "paidToName": { "type": "string" },
          "reference": { "type": "string" },
          "remarks": { "type": "string", "nullable": true },
          "isVoid": { "type": "boolean" },
          "createdAt": { "type": "string" }
        }
      },
      "TransactionCreate": {
        "type": "object",
        "properties": {
          "batchId": { "type": "string" },
          "date": { "type": "string" },
          "building": { "type": "string" },
          "fundingNature": { "type": "string" },
          "category": { "type": "string" },
          "description": { "type": "string" },
          "quantity": { "type": "number" },
          "unitCost": { "type": "number" },
          "amount": { "type": "number" },
          "paidBy": { "type": "string" },
          "paidTo": { "type": "string" },
          "reference": { "type": "string" },
          "remarks": { "type": "string" },
          "type": { "type": "string" },
          "feedItemId": { "type": "integer" }
        },
        "required": ["date", "fundingNature", "category", "description"]
      },
...
    }
  }
}